move bulk of code out of management command
