+++ corehq/apps/dump_reload/exceptions.py
@@ -1,0 +1,2 @@
+class DomainDumpError(Exception):
+    pass
